// src/app/models/Models.ts
export class Administrador implements IAdministrador {
  contrasena?: string | undefined;
  correo?: string | undefined;

  constructor(data?: IAdministrador) {
      if (data) {
          for (var property in data) {
              if (data.hasOwnProperty(property))
                  (<any>this)[property] = (<any>data)[property];
          }
      }
  }

  init(_data?: any) {
      if (_data) {
          this.contrasena = _data["Contrasena"];
          this.correo = _data["Correo"];
      }
  }

  static fromJS(data: any): Administrador {
      data = typeof data === 'object' ? data : {};
      let result = new Administrador();
      result.init(data);
      return result;
  }

  toJSON(data?: any) {
      data = typeof data === 'object' ? data : {};
      data["Contrasena"] = this.contrasena;
      data["Correo"] = this.correo;
      return data;
  }
}

export interface IAdministrador {
  contrasena?: string | undefined;
  correo?: string | undefined;
}
// ...
export class Usuario implements IUsuario {
  correoelectronico?: string | undefined;
  contrasena?: string | undefined;
  primernombre?: string | undefined;
  segundonombre?: string | undefined;
  primerapellido?: string | undefined;
  segundoapellido?: string | undefined;
  usuarioUbicacion?: UsuarioUbicacion | undefined;
  aposentos?: Aposento[] | undefined;
  certificadoGarantia?: CertificadoGarantium[] | undefined;
  dispositivos?: Dispositivo[] | undefined;
  duenoPasados?: DuenoPasado[] | undefined;
  pedidos?: Pedido[] | undefined;
  usuarioDireccions?: UsuarioDireccion[] | undefined;

  constructor(data?: IUsuario) {
      if (data) {
          for (var property in data) {
              if (data.hasOwnProperty(property))
                  (<any>this)[property] = (<any>data)[property];
          }
      }
  }

  init(_data?: any) {
      if (_data) {
          this.correoelectronico = _data["Correoelectronico"];
          this.contrasena = _data["Contrasena"];
          this.primernombre = _data["Primernombre"];
          this.segundonombre = _data["Segundonombre"];
          this.primerapellido = _data["Primerapellido"];
          this.segundoapellido = _data["Segundoapellido"];
          this.usuarioUbicacion = _data["UsuarioUbicacion"] ? UsuarioUbicacion.fromJS(_data["UsuarioUbicacion"]) : <any>undefined;
          if (Array.isArray(_data["Aposentos"])) {
              this.aposentos = [] as any;
              for (let item of _data["Aposentos"])
                  this.aposentos!.push(Aposento.fromJS(item));
          }
          if (Array.isArray(_data["CertificadoGarantia"])) {
              this.certificadoGarantia = [] as any;
              for (let item of _data["CertificadoGarantia"])
                  this.certificadoGarantia!.push(CertificadoGarantium.fromJS(item));
          }
          if (Array.isArray(_data["Dispositivos"])) {
              this.dispositivos = [] as any;
              for (let item of _data["Dispositivos"])
                  this.dispositivos!.push(Dispositivo.fromJS(item));
          }
          if (Array.isArray(_data["DuenoPasados"])) {
              this.duenoPasados = [] as any;
              for (let item of _data["DuenoPasados"])
                  this.duenoPasados!.push(DuenoPasado.fromJS(item));
          }
          if (Array.isArray(_data["Pedidos"])) {
              this.pedidos = [] as any;
              for (let item of _data["Pedidos"])
                  this.pedidos!.push(Pedido.fromJS(item));
          }
          if (Array.isArray(_data["UsuarioDireccions"])) {
              this.usuarioDireccions = [] as any;
              for (let item of _data["UsuarioDireccions"])
                  this.usuarioDireccions!.push(UsuarioDireccion.fromJS(item));
          }
      }
  }

  static fromJS(data: any): Usuario {
      data = typeof data === 'object' ? data : {};
      let result = new Usuario();
      result.init(data);
      return result;
  }

  toJSON(data?: any) {
      data = typeof data === 'object' ? data : {};
      data["Correoelectronico"] = this.correoelectronico;
      data["Contrasena"] = this.contrasena;
      data["Primernombre"] = this.primernombre;
      data["Segundonombre"] = this.segundonombre;
      data["Primerapellido"] = this.primerapellido;
      data["Segundoapellido"] = this.segundoapellido;
      data["UsuarioUbicacion"] = this.usuarioUbicacion ? this.usuarioUbicacion.toJSON() : <any>undefined;
      if (Array.isArray(this.aposentos)) {
          data["Aposentos"] = [];
          for (let item of this.aposentos)
              data["Aposentos"].push(item.toJSON());
      }
      if (Array.isArray(this.certificadoGarantia)) {
          data["CertificadoGarantia"] = [];
          for (let item of this.certificadoGarantia)
              data["CertificadoGarantia"].push(item.toJSON());
      }
      if (Array.isArray(this.dispositivos)) {
          data["Dispositivos"] = [];
          for (let item of this.dispositivos)
              data["Dispositivos"].push(item.toJSON());
      }
      if (Array.isArray(this.duenoPasados)) {
          data["DuenoPasados"] = [];
          for (let item of this.duenoPasados)
              data["DuenoPasados"].push(item.toJSON());
      }
      if (Array.isArray(this.pedidos)) {
          data["Pedidos"] = [];
          for (let item of this.pedidos)
              data["Pedidos"].push(item.toJSON());
      }
      if (Array.isArray(this.usuarioDireccions)) {
          data["UsuarioDireccions"] = [];
          for (let item of this.usuarioDireccions)
              data["UsuarioDireccions"].push(item.toJSON());
      }
      return data;
  }
}
// ...
export class DistribuidorUbicacion implements IDistribuidorUbicacion {
  cedulajuridicadistribuidor?: string | undefined;
  paisregion?: string | undefined;
  cedulajuridicadistribuidorNavigation?: Distribuidor | undefined;
  paisregionNavigation?: Region | undefined;

  constructor(data?: IDistribuidorUbicacion) {
      if (data) {
          for (var property in data) {
              if (data.hasOwnProperty(property))
                  (<any>this)[property] = (<any>data)[property];
          }
      }
  }

  init(_data?: any) {
      if (_data) {
          this.cedulajuridicadistribuidor = _data["Cedulajuridicadistribuidor"];
          this.paisregion = _data["Paisregion"];
          this.cedulajuridicadistribuidorNavigation = _data["CedulajuridicadistribuidorNavigation"] ? Distribuidor.fromJS(_data["CedulajuridicadistribuidorNavigation"]) : <any>undefined;
          this.paisregionNavigation = _data["PaisregionNavigation"] ? Region.fromJS(_data["PaisregionNavigation"]) : <any>undefined;
      }
  }

  static fromJS(data: any): DistribuidorUbicacion {
      data = typeof data === 'object' ? data : {};
      let result = new DistribuidorUbicacion();
      result.init(data);
      return result;
  }

  toJSON(data?: any) {
      data = typeof data === 'object' ? data : {};
      data["Cedulajuridicadistribuidor"] = this.cedulajuridicadistribuidor;
      data["Paisregion"] = this.paisregion;
      data["CedulajuridicadistribuidorNavigation"] = this.cedulajuridicadistribuidorNavigation ? this.cedulajuridicadistribuidorNavigation.toJSON() : <any>undefined;
      data["PaisregionNavigation"] = this.paisregionNavigation ? this.paisregionNavigation.toJSON() : <any>undefined;
      return data;
  }
}

export interface IDistribuidorUbicacion {
  cedulajuridicadistribuidor?: string | undefined;
  paisregion?: string | undefined;
  cedulajuridicadistribuidorNavigation?: Distribuidor | undefined;
  paisregionNavigation?: Region | undefined;
}

export class Distribuidor implements IDistribuidor {
  cedulajuridica?: string | undefined;
  nombre?: string | undefined;
  distribuidorUbicacion?: DistribuidorUbicacion | undefined;
  dispositivos?: Dispositivo[] | undefined;

  constructor(data?: IDistribuidor) {
      if (data) {
          for (var property in data) {
              if (data.hasOwnProperty(property))
                  (<any>this)[property] = (<any>data)[property];
          }
      }
  }

  init(_data?: any) {
      if (_data) {
          this.cedulajuridica = _data["Cedulajuridica"];
          this.nombre = _data["Nombre"];
          this.distribuidorUbicacion = _data["DistribuidorUbicacion"] ? DistribuidorUbicacion.fromJS(_data["DistribuidorUbicacion"]) : <any>undefined;
          if (Array.isArray(_data["Dispositivos"])) {
              this.dispositivos = [] as any;
              for (let item of _data["Dispositivos"])
                  this.dispositivos!.push(Dispositivo.fromJS(item));
          }
      }
  }

  static fromJS(data: any): Distribuidor {
      data = typeof data === 'object' ? data : {};
      let result = new Distribuidor();
      result.init(data);
      return result;
  }

  toJSON(data?: any) {
      data = typeof data === 'object' ? data : {};
      data["Cedulajuridica"] = this.cedulajuridica;
      data["Nombre"] = this.nombre;
      data["DistribuidorUbicacion"] = this.distribuidorUbicacion ? this.distribuidorUbicacion.toJSON() : <any>undefined;
      if (Array.isArray(this.dispositivos)) {
          data["Dispositivos"] = [];
          for (let item of this.dispositivos)
              data["Dispositivos"].push(item.toJSON());
      }
      return data;
  }
}

export interface IDistribuidor {
  cedulajuridica?: string | undefined;
  nombre?: string | undefined;
  distribuidorUbicacion?: DistribuidorUbicacion | undefined;
  dispositivos?: Dispositivo[] | undefined;
}
// ...
export class TipoDispositivo implements ITipoDispositivo {
  nombre?: string | undefined;
  descripcion?: string | undefined;
  tiempogarantia!: number;
  imagen?: string | undefined;
  certificadoGarantia?: CertificadoGarantium[] | undefined;
  dispositivos?: Dispositivo[] | undefined;

  constructor(data?: ITipoDispositivo) {
      if (data) {
          for (var property in data) {
              if (data.hasOwnProperty(property))
                  (<any>this)[property] = (<any>data)[property];
          }
      }
  }

  init(_data?: any) {
      if (_data) {
          this.nombre = _data["Nombre"];
          this.descripcion = _data["Descripcion"];
          this.tiempogarantia = _data["Tiempogarantia"];
          this.imagen = _data["Imagen"];
          if (Array.isArray(_data["CertificadoGarantia"])) {
              this.certificadoGarantia = [] as any;
              for (let item of _data["CertificadoGarantia"])
                  this.certificadoGarantia!.push(CertificadoGarantium.fromJS(item));
          }
          if (Array.isArray(_data["Dispositivos"])) {
              this.dispositivos = [] as any;
              for (let item of _data["Dispositivos"])
                  this.dispositivos!.push(Dispositivo.fromJS(item));
          }
      }
  }

  static fromJS(data: any): TipoDispositivo {
      data = typeof data === 'object' ? data : {};
      let result = new TipoDispositivo();
      result.init(data);
      return result;
  }

  toJSON(data?: any) {
      data = typeof data === 'object' ? data : {};
      data["Nombre"] = this.nombre;
      data["Descripcion"] = this.descripcion;
      data["Tiempogarantia"] = this.tiempogarantia;
      data["Imagen"] = this.imagen;
      if (Array.isArray(this.certificadoGarantia)) {
          data["CertificadoGarantia"] = [];
          for (let item of this.certificadoGarantia)
              data["CertificadoGarantia"].push(item.toJSON());
      }
      if (Array.isArray(this.dispositivos)) {
          data["Dispositivos"] = [];
          for (let item of this.dispositivos)
              data["Dispositivos"].push(item.toJSON());
      }
      return data;
  }
}

export interface ITipoDispositivo {
  nombre?: string | undefined;
  descripcion?: string | undefined;
  tiempogarantia: number;
  imagen?: string | undefined;
  certificadoGarantia?: CertificadoGarantium[] | undefined;
  dispositivos?: Dispositivo[] | undefined;
}