// src/app/models/Clase.ts
export class Clase {
    DatetimeInicio: Date|undefined;
    DatetimeFinal: Date|undefined;
    Tipo: string|undefined;
    Modalidad: string|undefined;
    CedulaEmpleado: string|undefined;
    ImpartidoEnSucursal: string|undefined;
    Imagen?: string|undefined;
    Identificador: number|undefined;
    CupoDisponible: number|undefined;
    constructor() {
    }
  }